Replace ServiceEntrySpec with ServiceEntryTests.
